+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `Distinct<T>()` constraint for asserting pairwise distinctness of expressions
+
 ## [0.0.7] - 2025-12-10
 
 ### Added